requirements.txt
torch
torchvision
matplotlib
numpy
scikit-learn
einops
termcolor
kornia
pillow
tqdm
